pyproject.toml
[tool.ruff]
line-length = 88
[tool.ruff.lint]
select = ["E", "F", "W", "C"]

[tool.mypy]
python_version = 3.11
strict = true

[project]
name = "caching-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "annotated-types (==0.7.0)",
    "anyio (==4.10.0)",
    "certifi (==2025.8.3)",
    "charset-normalizer (==3.4.3)",
    "click (==8.2.1)",
    "colorama (==0.4.6)",
    "fastapi (==0.116.1)",
    "greenlet (==3.2.4)",
    "h11 (==0.16.0)",
    "httptools (==0.6.4)",
    "idna (==3.10)",
    "mypy (==1.18.2)",
    "mypy-extensions (==1.1.0)",
    "pathspec (==0.12.1)",
    "psycopg2-binary (==2.9.10)",
    "pydantic (==2.11.7)",
    "pydantic-core (==2.33.2)",
    "python-dotenv (==1.1.1)",
    "pyyaml (==6.0.2)",
    "requests (==2.32.5)",
    "ruff (==0.13.1)",
    "sniffio (==1.3.1)",
    "sqlalchemy (==2.0.43)",
    "starlette (==0.47.3)",
    "types-greenlet (==3.2.0.20250915)",
    "types-pyyaml (==6.0.12.20250915)",
    "types-requests (==2.32.4.20250913)",
    "types-ujson (==5.10.0.20250822)",
    "typing-inspection (==0.4.1)",
    "typing-extensions (==4.15.0)",
    "urllib3 (==2.5.0)",
    "uvicorn (==0.35.0)",
    "watchfiles (==1.1.0)",
    "websockets (==15.0.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
